requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pytz
requests
pydantic
flask
gunicorn
pydantic-settings
httpx
python-multipart
motor
Jinja2
pydantic[email]